01_webapp/requirements.txt
datasets
sentencepiece
transformers==4.25.1
pytorch-lightning==1.6.4
timm==0.5.4
streamlit
donut-python
PyPDF2
pdf2image
Pillow
